open-iscsi: Fix installation of libopeniscsiusr symlinks
I discovered this issue by scanning the yum repo for "provides" that are
duplicated across different packages. The scan flagged open-iscsi-lib and
open-iscsi-dev as both providing "libopeniscsiusr.so.0.1.0()(64bit)".

This led to the discovery of how the open-iscsi build was using install(1) to
install symlinks, but install(1) will dereference the symlinks when copying to
the target. So I've added a patch to create the proper symlinks during
installation instead.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>